# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for CandyPro OEM candy manufacturing website",
        "title": "CandyPro OEM API",
        "contact": {
            "name": "API Support",
            "email": "info@candypro.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/products/{id}": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-products"
                ],
                "summary": "Admin update product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-products"
                ],
                "summary": "Admin delete product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
...
}